requirements.txt
asgiref==3.6.0
black==23.1.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
coverage==6.5.0
coveralls==3.3.1
Django==2.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
docopt==0.6.2
flake8==5.0.4
idna==3.4
importlib-metadata==4.2.0
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.0
pycodestyle==2.9.1
pyflakes==2.5.0
PyJWT==2.6.0
python-dotenv==0.21.1
pytz==2022.7.1
random-username==1.0.2
requests==2.28.2
sqlparse==0.4.3
tomli==2.0.1
typed-ast==1.5.4
typing_extensions==4.5.0
urllib3==1.26.14
zipp==3.15.0
